Change branch option name
Because heroku now supports both `master` and `main` target branch names, the meaning of the `branch` option in this action is ambiguous.

I decided to remove the check that current branch is identical to the onen specified, as it is already something that should be checked inside the workflow CI file. Hence the branch specified by the option is the name of the target heroku branch (that is, either `master` or `main`).